Elminate meaningless typecast

+++ src/sbbs3/useredit.cpp
@@ -517,7 +517,7 @@ void sbbs_t::useredit(int usernumber)
 				if (!noyes(text[UeditCopyUserQ])) {
 					bputs(text[UeditCopyUserToSlot]);
 					i = getnum(lastuser(&cfg));
-					if ((int)i > 0) {
+					if (i > 0) {
 						user.number = i;
 						putusername(&cfg, user.number, user.alias);
 						putuserdat(&user);
